+++ src/lib/models/reflections/abstract.ts
@@ -314,8 +314,6 @@ export abstract class Reflection {
 
     /**
      * Url safe alias for this reflection.
-     *
-     * @see {@link BaseReflection.getAlias}
      */
     private _alias?: string;