+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Fixed infinite loop caused by a fix for some complicated union/intersection types, #2468.
+-   Improved infinite loop detection in type converter to reduce false positives.
 
 ## v0.25.5 (2024-01-01)