+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+## v0.25.6 (2024-01-01)
+
 ### Bug Fixes
 
 -   Fixed infinite loop caused by a fix for some complicated union/intersection types, #2468.